moved dupe functions to util
